requirements.txt
google-cloud-bigquery==2.34.2
honeygain.py>=1.1,<2.0